lw3/requirements.txt
apache-airflow==2.7.3
pandas==2.1.3
numpy==1.26.2
requests==2.31.0
pathlib==1.0.1